base ON snowCat Framework
this demo is based on SnowCat Framework ,which is built by Moky ,who is
very busy and have no time to give a Demo. But he has a dream that he
want to make programming is simple as making net page. so he made
snowCat framework. I am very moving on by his words. this simple demo
is based on his tutor.